Port CVE-2010-4538 forward
This is a vulnerability that allows a buffer overflow in the initialized
data segment area of the target executable and can be used for a DoS
attack and could potentially be used to alter the code execution. This
demo the exploit packet causes tshark to die on a SIGBUS signal. The
code change that fixed this vulnerability can be found at:

http://anonsvn.wireshark.org/viewvc/trunk/epan/dissectors/packet-enttec.c?r1=35318&r2=35317&pathrev=35318