Port CVE-2013-5719 forward
This is a vulnerability that allows a specially crafted ASSA
Abloy R3 packet to cause Wireshark to enter an infinite loop. In
dissect_r3_upstreamfgfield_checkresults():

        for (i = 0; i < len; i += tvb_get_guint8 (tvb, i))...

if tvb_get_guint8() ever encounters a 0, i never gets incremented and
the code loops forever. The code change that fixed this vulnerbility can
be found at:

http://anonsvn.wireshark.org/viewvc/trunk/epan/dissectors/packet-assa_r3.c?r1=51196&r2=51195&pathrev=51196