Port CVE-2011-1591 forward
This is stack based buffer overflow in the DECT dissector code that can
be used to execute arbitrary code on the target system. For this demo it
is assumed that stack protection (such as the "-fstack_protect" compiler
option) and Address Space Layout Randomization (ASLR) were not enabled
when wireshark/tshark was compiled and executed. Please note that this
is the default case for the FreeBSD wireshark 1.10.2 port. It has been
demonstrated that stack protection (i.e. Microsoft's Data Execution
Prevention/Safe Structured Exception Handling or DEP/SafeSEH) and ASLR
can be bypassed using a more complicated exploit with this vulnerability
(see http://www.exploit-db.com/exploits/17186/).